Enhance action lifecycle and layer ownership: update macro sender for owned-keycode dispatch and override modifier registration for shared ownership model

+++ users/noah/lib/state/layer_ownership.c
@@ -20,20 +20,20 @@ static inline bool layer_ownership_keypos_equal(keypos_t lhs, keypos_t rhs) {
     return lhs.row == rhs.row && lhs.col == rhs.col;
 }
 
-static int8_t layer_ownership_find_slot_for_key(keypos_t key_pos) {
-    for (uint8_t i = 0; i < ARRAY_SIZE(layer_momentary_bindings); i++) {
+static int16_t layer_ownership_find_slot_for_key(keypos_t key_pos) {
+    for (uint16_t i = 0; i < ARRAY_SIZE(layer_momentary_bindings); i++) {
         if (layer_momentary_bindings[i].active && layer_ownership_keypos_equal(layer_momentary_bindings[i].key_pos, key_pos)) {
-            return (int8_t)i;
+            return (int16_t)i;
         }
     }
 
     return -1;
 }
 
-static int8_t layer_ownership_find_free_slot(void) {
-    for (uint8_t i = 0; i < ARRAY_SIZE(layer_momentary_bindings); i++) {
+static int16_t layer_ownership_find_free_slot(void) {
+    for (uint16_t i = 0; i < ARRAY_SIZE(layer_momentary_bindings); i++) {
         if (!layer_momentary_bindings[i].active) {
-            return (int8_t)i;
+            return (int16_t)i;
         }
     }
 
@@ -65,7 +65,7 @@ static bool layer_ownership_apply_layer(uint8_t layer) {
     return false;
 }
 
-static bool layer_ownership_remove_slot(uint8_t slot) {
+static bool layer_ownership_remove_slot(uint16_t slot) {
     if (slot >= ARRAY_SIZE(layer_momentary_bindings) || !layer_momentary_bindings[slot].active) {
         return false;
     }
@@ -127,14 +127,14 @@ void layer_ownership_momentary_press(keypos_t key_pos, uint8_t layer) {
         return;
     }
 
-    int8_t slot = layer_ownership_find_slot_for_key(key_pos);
+    int16_t slot = layer_ownership_find_slot_for_key(key_pos);
     if (slot >= 0) {
         if (layer_momentary_bindings[slot].layer == layer) {
             layer_ownership_apply_layer(layer);
             return;
         }
 
-        layer_ownership_remove_slot((uint8_t)slot);
+        layer_ownership_remove_slot((uint16_t)slot);
     } else {
         slot = layer_ownership_find_free_slot();
         if (slot < 0) {
@@ -154,11 +154,11 @@ void layer_ownership_momentary_press(keypos_t key_pos, uint8_t layer) {
 }
 
 bool layer_ownership_momentary_release(keypos_t key_pos) {
-    int8_t slot = layer_ownership_find_slot_for_key(key_pos);
+    int16_t slot = layer_ownership_find_slot_for_key(key_pos);
 
     if (slot < 0) {
         return false;
     }
 
-    return layer_ownership_remove_slot((uint8_t)slot);
+    return layer_ownership_remove_slot((uint16_t)slot);
 }
